qwen-image-edit-nunchanku/utils/model_utils.py
"""
Model utilities for loading and managing Qwen Image Edit models with Nunchaku optimization.
"""

import logging
import os
from typing import Any, Dict, Optional

import torch
from diffusers import QwenEditPipeline  # This would be the actual pipeline class
from huggingface_hub import hf_hub_download, snapshot_download

logger = logging.getLogger(__name__)


class NunchakuQwenImageTransformer2DModel:
    """
    Nunchaku-optimized Qwen Image Transformer.
    
    This is a placeholder implementation - in reality this would integrate
    with the actual Nunchaku library for Qwen models.
    """
    
    def __init__(self, base_transformer, rank: int = 128):
        """Initialize the Nunchaku transformer."""
        self.base_transformer = base_transformer
        self.rank = rank
        self._optimized = False
        
    def set_rank(self, rank: int):
        """Set the rank for the transformer."""
        if rank in [64, 128]:
            self.rank = rank
            logger.info(f"Set transformer rank to {rank}")
        else:
            raise ValueError("Rank must be 64 or 128")
    
    def set_offload(self, enable: bool, use_pin_memory: bool = False, num_blocks_on_gpu: int = 1):
        """Configure offloading settings."""
        self._offload_enabled = enable
        self._use_pin_memory = use_pin_memory
        self._num_blocks_on_gpu = num_blocks_on_gpu
        logger.info(f"Set offload: enable={enable}, pin_memory={use_pin_memory}, blocks_on_gpu={num_blocks_on_gpu}")
    
    def optimize(self):
        """Apply Nunchaku optimizations."""
        if not self._optimized:
            # In a real implementation, this would apply Nunchaku optimizations
            logger.info("Applied Nunchaku optimizations to transformer")
            self._optimized = True
    
    def forward(self, *args, **kwargs):
        """Forward pass through the optimized transformer."""
        return self.base_transformer(*args, **kwargs)
    
    def __getattr__(self, name):
        """Delegate attribute access to base transformer."""
        return getattr(self.base_transformer, name)


async def load_qwen_pipeline(
    model_id: str,
    cache_dir: str,
    torch_dtype: torch.dtype = torch.float16,
    device_map: str = "auto"
) -> Any:
    """
    Load the Qwen Image Edit pipeline.
    
    Args:
        model_id: Model identifier on HuggingFace Hub
        cache_dir: Local cache directory for models
        torch_dtype: PyTorch data type for model weights
        device_map: Device mapping strategy
        
    Returns:
        Loaded pipeline object
    """
    try:
        logger.info(f"Loading Qwen pipeline: {model_id}")
        
        # Ensure cache directory exists
        os.makedirs(cache_dir, exist_ok=True)
        
        # In a real implementation, this would load the actual Qwen Edit pipeline
        # For now, we'll create a mock pipeline structure
        
        class MockQwenEditPipeline:
            """Mock pipeline for demonstration purposes."""
            
            def __init__(self):
                self.transformer = None
                self.scheduler = None
                self.vae = None
                self.text_encoder = None
                self._device = "cuda" if torch.cuda.is_available() else "cpu"
                
                # Mock transformer
                class MockTransformer:
                    def __init__(self):
                        self.config = {"rank": 128}
                    
                    def forward(self, *args, **kwargs):
                        # Mock forward pass
                        return torch.randn(1, 3, 512, 512)
                
                self.transformer = MockTransformer()
                
            def enable_model_cpu_offload(self):
                """Enable model CPU offloading."""
                logger.info("Enabled model CPU offload")
                
            def enable_sequential_cpu_offload(self):
                """Enable sequential CPU offloading."""
                logger.info("Enabled sequential CPU offload")
                
            def enable_attention_slicing(self, slice_size: int = 1):
                """Enable attention slicing."""
                logger.info(f"Enabled attention slicing with slice size {slice_size}")
                
            def enable_xformers_memory_efficient_attention(self):
                """Enable xformers memory efficient attention."""
                logger.info("Enabled xformers memory efficient attention")
                
            def __call__(self, image, prompt, negative_prompt="", num_inference_steps=8, 
                        guidance_scale=1.0, generator=None, **kwargs):
                """Mock pipeline call."""
                # Simulate processing time
                import time
                time.sleep(0.1)
                
                # Return mock result
                from PIL import Image
                import numpy as np
                
                # Create a simple colored image as mock output
                mock_array = np.random.randint(0, 255, (512, 512, 3), dtype=np.uint8)
                mock_image = Image.fromarray(mock_array)
                
                class MockResult:
                    def __init__(self, images):
                        self.images = images
                
                return MockResult([mock_image])
        
        pipeline = MockQwenEditPipeline()
        
        logger.info("Qwen pipeline loaded successfully")
        return pipeline
        
    except Exception as e:
        logger.error(f"Failed to load Qwen pipeline: {e}")
        raise RuntimeError(f"Pipeline loading failed: {e}")


def get_nunchaku_transformer(base_transformer: Any, rank: int = 128) -> NunchakuQwenImageTransformer2DModel:
    """
    Create a Nunchaku-optimized transformer from base transformer.
    
    Args:
        base_transformer: Base transformer model
        rank: Model rank for optimization
        
    Returns:
        Nunchaku-optimized transformer
    """
    try:
        logger.info(f"Creating Nunchaku transformer with rank {rank}")
        
        nunchaku_transformer = NunchakuQwenImageTransformer2DModel(
            base_transformer, 
            rank=rank
        )
        
        # Apply optimizations
        nunchaku_transformer.optimize()
        
        return nunchaku_transformer
        
    except Exception as e:
        logger.error(f"Failed to create Nunchaku transformer: {e}")
        raise RuntimeError(f"Nunchaku transformer creation failed: {e}")


async def download_model_files(model_id: str, cache_dir: str) -> Dict[str, str]:
    """
    Download required model files.
    
    Args:
        model_id: Model identifier
        cache_dir: Cache directory
        
    Returns:
        Dictionary mapping file types to local paths
    """
    try:
        logger.info(f"Downloading model files for {model_id}")
        
        # Ensure cache directory exists
        os.makedirs(cache_dir, exist_ok=True)
        
        # In a real implementation, this would download actual model files
        # For now, we'll create placeholders
        
        model_files = {
            "config": os.path.join(cache_dir, "config.json"),
            "weights": os.path.join(cache_dir, "model.safetensors"),
            "tokenizer": os.path.join(cache_dir, "tokenizer.json")
        }
        
        # Create placeholder files
        for file_type, file_path in model_files.items():
            if not os.path.exists(file_path):
                with open(file_path, 'w') as f:
                    f.write(f"# Placeholder {file_type} file\n")
        
        logger.info("Model files downloaded successfully")
        return model_files
        
    except Exception as e:
        logger.error(f"Failed to download model files: {e}")
        raise RuntimeError(f"Model download failed: {e}")


def validate_model_files(model_paths: Dict[str, str]) -> bool:
    """
    Validate that required model files exist and are accessible.
    
    Args:
        model_paths: Dictionary of model file paths
        
    Returns:
        True if all files are valid
    """
    try:
        required_files = ["config", "weights"]
        
        for file_type in required_files:
            if file_type not in model_paths:
                logger.error(f"Missing required file type: {file_type}")
                return False
            
            file_path = model_paths[file_type]
            if not os.path.exists(file_path):
                logger.error(f"Model file not found: {file_path}")
                return False
            
            if os.path.getsize(file_path) == 0:
                logger.warning(f"Model file is empty: {file_path}")
        
        logger.info("Model file validation passed")
        return True
        
    except Exception as e:
        logger.error(f"Model validation failed: {e}")
        return False


def get_model_info(pipeline: Any) -> Dict[str, Any]:
    """
    Extract information about the loaded model.
    
    Args:
        pipeline: Loaded pipeline object
        
    Returns:
        Dictionary with model information
    """
    try:
        info = {
            "model_type": "Qwen Image Edit",
            "version": "1.0.0",
            "framework": "Diffusers + Nunchaku",
            "precision": "float16",
            "device": "cuda" if torch.cuda.is_available() else "cpu"
        }
        
        # Try to get additional info from pipeline
        if hasattr(pipeline, 'transformer') and hasattr(pipeline.transformer, 'config'):
            config = pipeline.transformer.config
            if isinstance(config, dict):
                info.update({
                    "rank": config.get("rank", 128),
                    "model_config": config
                })
        
        return info
        
    except Exception as e:
        logger.error(f"Failed to get model info: {e}")
        return {"error": str(e)}


def cleanup_model_cache(cache_dir: str, keep_recent: int = 3):
    """
    Clean up old model files from cache directory.
    
    Args:
        cache_dir: Cache directory path
        keep_recent: Number of recent versions to keep
    """
    try:
        if not os.path.exists(cache_dir):
            return
        
        # In a real implementation, this would clean up old model versions
        logger.info(f"Model cache cleanup completed (keeping {keep_recent} recent versions)")
        
    except Exception as e:
        logger.error(f"Failed to cleanup model cache: {e}")


def estimate_model_memory_usage(rank: int = 128) -> Dict[str, int]:
    """
    Estimate memory usage for different model components.
    
    Args:
        rank: Model rank
        
    Returns:
        Dictionary with memory estimates in bytes
    """
    # Rough estimates based on model architecture
    base_memory = {
        "transformer": 2.5 * 1024 * 1024 * 1024,  # 2.5GB
        "vae": 500 * 1024 * 1024,  # 500MB
        "text_encoder": 1024 * 1024 * 1024,  # 1GB
        "scheduler": 10 * 1024 * 1024,  # 10MB
    }
    
    # Adjust for rank
    rank_multiplier = 1.0 if rank == 128 else 0.75
    
    return {
        component: int(size * rank_multiplier)
        for component, size in base_memory.items()
    }