qwen-image-edit-nunchanku/api_requirements.txt
# FastAPI and web server dependencies
fastapi[standard]==0.115.12
uvicorn[standard]==0.24.0
pydantic==2.11.4
python-multipart==0.0.6

# PyTorch and ML dependencies
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
xformers>=0.0.29

# Diffusers and Hugging Face
diffusers>=0.30.0
transformers>=4.53.0
accelerator>=1.8.1
huggingface-hub[hf_transfer]>=0.33.1
safetensors>=0.5.3
sentencepiece>=0.2.0
para-attn>=0.3.32

# Image processing
Pillow>=11.2.1
opencv-python-headless>=4.11.0.86
numpy>=2.2.4

# HTTP client and utilities
requests>=2.31.0
aiohttp>=3.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Optional: Nunchaku optimization (replace with actual wheel URL)
# nunchaku @ https://github.com/nunchaku-tech/nunchaku/releases/download/v0.3.1/nunchaku-0.3.1+torch2.7-cp311-cp311-linux_x86_64.whl